+++ my-website/src/components/Contact/Contact.js
@@ -2,9 +2,12 @@ import React from "react";
 
 import "./Contact.css";
 
+import Navbar from "../Navbar/Navbar";
+
 function Contact(){
     return(
         <>
+            <Navbar />
             <h1>Contact Page!</h1>
         </>
     );
