+++ my-website/src/components/Fun-Stuff/Funstuff.js
@@ -2,13 +2,14 @@ import React from "react";
 
 import "./Funstuff.css";
 
-import Search from "../Search/Search";
+import Navbar from "../Navbar/Navbar";
 
 function Funstuff() {
     return(
         <>
+            <Navbar />
             <h1>Fun-Stuff Page!</h1>
-            <Search />
+            
         </>
     );
 }
