+++ my-website/src/components/Home/Home.css
@@ -2,15 +2,11 @@
     
 }
 
-.blob{
-    position: absolute;
-    z-index: 1;
 
-    border: 3px red solid;
-}
+
 
 .pos1 {
-    width: 25%;
+    width: 100%;
     height: 100%;
     
     top: 0;
@@ -34,11 +30,6 @@
 }
 
 
-.content {
-
-    position: absolute;
-    z-index: 2;
-}
 
 /*Tablet min-width*/
 @media all and (min-width: 700px) {
