+++ my-website/src/components/Navbar/Navbar.css
@@ -11,6 +11,7 @@ nav {
 }
 
 .navbar {
+  
     
 }
